Add files via upload
GameManager spawns enemies at random points on the game. It's not yet linked to day/night cycle and instead uses a standard TD wave system for now.

Enemy movement also modified to dynamically find the player's base on it's own as long as the game object is named "Home Base" in the unity editor.